add filtered csv file
The csv given from the database
https://domo-support.domo.com/s/article/360043931814?language=en_US
(under 200,000+ Jeopardy Questions header)
was filtered in this manner:
The fields are Value, Category, Answer, Question, with each field
delimited by pipes.
The questions contain no links to media. Only text.
The values of each question are fitted to 200, 400, 600, 800, and 1000,
in that order.
The questions are grouped by category, and are in groups of 5.
All double jeopardy questions are removed.